fix Desktop di
